% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationarity.R
\name{adf_test}
\alias{adf_test}
\title{Run an Augmented Dickey-Fuller Test}
\usage{
adf_test(x, alternative = "stationary", ..., .alpha = 0.05)
}
\arguments{
\item{x}{a numeric vector or time series.}

\item{alternative}{indicates the alternative hypothesis and must be
    one of \code{"stationary"} (default) or \code{"explosive"}. You can
    specify just the initial letter.}

\item{...}{further arguments passed to \code{\link[tseries]{adf.test}}.}

\item{.alpha}{critical p-value used to determine test conclusion.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
A wrapper around \code{\link[tseries]{adf.test}} that standardizes the
inputs and outputs.
}
\examples{
set.seed(123)
x <- runif(n = 1000)

adf_test(x)
adf_test(x, alternative = "explosive")

}
