% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barchart.R
\name{barchart}
\alias{barchart}
\title{Draw a bar chart}
\usage{
barchart(df, x, y)
}
\arguments{
\item{df}{A dataframe}

\item{x}{Variable to divide up bars by}

\item{y}{Variable that gives height of each var}
}
\description{
Draw a bar chart
}
\examples{
barchart(ggplot2::diamonds, cut)
barchart(mtcars10, cyl)
}
