% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg-plots.R
\name{calc_pred_vs_resid}
\alias{calc_pred_vs_resid}
\alias{calc_pred_vs_resid.lm}
\title{Calculate Predictions and Residuals From a Model}
\usage{
calc_pred_vs_resid(object, data)

\method{calc_pred_vs_resid}{lm}(object, data)
}
\arguments{
\item{object}{A model object (such as a fitted \code{\link[stats:lm]{lm}} object).}

\item{data}{(Optional) a data set used to create predictions (if omitted,
the fitted values are used).}
}
\value{
A \link[tibble:tibble-package]{tibble} with columns \code{.pred} and \code{.resid}.
}
\description{
From a model, calculate the predicted values from a set of data and then
calculate the residuals of those predictions.
}
\examples{
library(dplyr)
library(parsnip)
library(tidytest)

#> `lm` Method
mod_lm_fit <- lm(mpg ~ disp + wt + hp, data = mtcars)

calc_pred_vs_resid(mod_lm_fit)
calc_pred_vs_resid(mod_lm_fit, mtcars[1:15, ])

}
