% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tseries-acf.R
\name{calculate_acf}
\alias{calculate_acf}
\alias{calculate_acf.numeric}
\alias{calculate_pacf}
\alias{calculate_pacf.numeric}
\title{Calculate Autocorrelation Function (ACF) or Partial Autocorrelation Function
(PACF) Values}
\usage{
calculate_acf(x, .lags = NULL, .conf = 0.95)

\method{calculate_acf}{numeric}(x, .lags = NULL, .conf = 0.95)

calculate_pacf(x, .lags = NULL, .conf = 0.95)

\method{calculate_pacf}{numeric}(x, .lags = NULL, .conf = 0.95)
}
\arguments{
\item{x}{A numeric vector.}

\item{.lags}{(Optional) Maximum lag value used in calculations. If omitted,
will be limited to one less than the number of observations in \code{x}.}

\item{.conf}{(Optional) Confidence interval coverage probability. Default is
for a 95\% confidence interval.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Wrappers around \code{\link[stats:acf]{stats::acf()}} and \code{\link[stats:acf]{stats::pacf()}} that standardize the
inputs and outputs.
}
\examples{
library(tidytest)

set.seed(1914)
x <- rnorm(100)

#> ACF
calculate_acf(x)
calculate_acf(x, .lags = 10, .conf = 0.8)

#> PACF
calculate_pacf(x)
calculate_pacf(x, .lags = 10, .conf = 0.8)

}
\seealso{
\code{\link[=plot_acf]{plot_acf()}} \code{\link[=plot_pacf]{plot_pacf()}}
}
