% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{calculate_residuals}
\alias{calculate_residuals}
\alias{calculate_residuals.lm}
\title{Calculate Predictions and Residuals From a Model}
\usage{
calculate_residuals(object, data = NULL)

\method{calculate_residuals}{lm}(object, data = NULL)
}
\arguments{
\item{object}{A model object (such as a fitted \code{\link[stats:lm]{lm}} object).}

\item{data}{(Optional) a data set used to create predictions (if omitted,
the fitted values are used).}
}
\value{
A \link[tibble:tibble-package]{tibble} with columns \code{.pred} and \code{.resid}.
}
\description{
From a model, calculate the predicted values from a set of data and then
calculate the residuals of those predictions.
}
\details{
\subsection{Known Limitations}{

This function will not work if transformations to variables are done in the
model formula (\emph{e.g.}, \code{lm(mpg ~ wt + as.factor(vs), data = mtcars)}).
Variable transformations should be done to the data set before creating the
model formula.
}
}
\examples{
library(tidytest)

#> `lm` Method
mod_lm_fit <- lm(mpg ~ disp + wt + hp, data = mtcars)

calculate_residuals(mod_lm_fit)
calculate_residuals(mod_lm_fit, mtcars[1:15, ])

}
