% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg-multicollinearity.R
\name{calculate_vifs}
\alias{calculate_vifs}
\alias{calculate_vifs.default}
\alias{calculate_vifs.lm}
\alias{calculate_vifs._lm}
\title{Calculate Multicollinearity of Variables in a Linear Regression}
\usage{
calculate_vifs(object)

\method{calculate_vifs}{default}(object)

\method{calculate_vifs}{lm}(object)

\method{calculate_vifs}{`_lm`}(object)
}
\arguments{
\item{object}{A model object (such as a fitted \code{\link[stats:lm]{lm}} object).}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
A wrapper around \code{\link[car]{vif}} that standardizes the inputs and
outputs.
}
\examples{
library(dplyr)
library(parsnip)
library(tidytest)

#> `lm` Method
mod_lm_fit <- lm(mpg ~ disp + wt + hp, data = mtcars)

calculate_vifs(mod_lm_fit)

#> Tidymodels Method
mod_linreg_fit <- linear_reg() \%>\%
  set_engine("lm") \%>\%
  fit(mpg ~ disp + wt + hp, data = mtcars)

calculate_vifs(mod_linreg_fit)

}
