% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyfuns.R
\name{cyl_sum}
\alias{cyl_sum}
\title{Average over each cylinder}
\usage{
cyl_sum(x)
}
\arguments{
\item{x}{Column to be averaged}
}
\value{
tibble
}
\description{
Average over each cylinder
}
\examples{
cyl_sum(wt)
}
