% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diamonds_summary2.R
\name{diamonds_summary2}
\alias{diamonds_summary2}
\title{Summarize the diamonds dataset}
\usage{
diamonds_summary2(bins)
}
\arguments{
\item{bins}{the number of bins}
}
\value{
a plot
}
\description{
Summarize the diamonds dataset
}
