% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tseries-stationarity.R
\name{dickey_fuller_test}
\alias{dickey_fuller_test}
\title{Run an Augmented Dickey-Fuller (ADF) Test}
\usage{
dickey_fuller_test(object, alternative = "stationary", ..., .alpha = 0.05)
}
\arguments{
\item{object}{a model object (such as a fitted \code{lm} object).}

\item{alternative}{indicates the alternative hypothesis and must be
    one of \code{"stationary"} (default) or \code{"explosive"}. You can
    specify just the initial letter.}

\item{...}{further arguments passed to \code{\link[tseries]{adf.test}}.}

\item{.alpha}{critical p-value used to determine test conclusion.}
}
\value{
a \link[tibble:tibble-package]{tibble}.
}
\description{
A wrapper around \code{\link[tseries]{adf.test}} that standardizes the
inputs and outputs.
}
\details{
The hypotheses for this test are:
\itemize{
\item Null: Series has a unit root
\item Alternative: Series is either stationary of explosive (user defined)
}
}
\examples{
library(dplyr)
library(parsnip)
library(tidytest)

mod_fit <- parsnip::linear_reg() \%>\%
  set_engine("lm") \%>\%
  fit(mpg ~ disp + wt + hp, data = mtcars)

dickey_fuller_test(mod_fit)
dickey_fuller_test(mod_fit, alternative = "stationary")
dickey_fuller_test(mod_fit, alternative = "explosive")

}
