% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg-independence.R
\name{durbin_watson_test}
\alias{durbin_watson_test}
\title{Run a Durbin-Watston Test}
\usage{
durbin_watson_test(object, alternative = "two.sided", ..., .alpha = 0.05)
}
\arguments{
\item{object}{a model object (such as a fitted \code{lm} object).}

\item{alternative}{a character string specifying the alternative hypothesis.}

\item{...}{further arguments passed to \code{\link[lmtest]{dwtest}}.}

\item{.alpha}{critical p-value used to determine test conclusion.}
}
\value{
a \link[tibble:tibble-package]{tibble}.
}
\description{
A wrapper around \code{\link[lmtest]{dwtest}} that standardizes the inputs
and outputs.
}
\details{
The hypotheses for this test are:
\itemize{
\item Null: No Autocorrelation
\item Alternative: Autocorrelation
}
}
\examples{
library(dplyr)
library(parsnip)
library(tidytest)

mod_fit <- parsnip::linear_reg() \%>\%
  set_engine("lm") \%>\%
  fit(mpg ~ disp + wt + hp, data = mtcars)

durbin_watson_test(mod_fit)
durbin_watson_test(mod_fit, alternative = "greater")
durbin_watson_test(mod_fit, alternative = "two.sided")

}
