% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg-homoscedasticity.R
\name{goldfeld_quandt_test}
\alias{goldfeld_quandt_test}
\alias{goldfeld_quandt_test.lm}
\alias{goldfeld_quandt_test._lm}
\alias{goldfeld_quandt_test._glm}
\title{Run a Goldfeld-Quandt Test}
\usage{
goldfeld_quandt_test(object, alternative = "two.sided", ..., .alpha = 0.05)

\method{goldfeld_quandt_test}{lm}(object, alternative = "two.sided", ..., .alpha = 0.05)

\method{goldfeld_quandt_test}{`_lm`}(object, alternative = "two.sided", ..., .alpha = 0.05)

\method{goldfeld_quandt_test}{`_glm`}(object, alternative = "two.sided", ..., .alpha = 0.05)
}
\arguments{
\item{object}{A model object (such as a fitted \code{\link[stats:lm]{lm}} object).}

\item{alternative}{Passed to the \code{alternative} argument of \code{\link[lmtest:gqtest]{lmtest::gqtest()}}
(one of "two.sided", "greater", or "less". Default is "two.sided").}

\item{...}{Further arguments passed to \code{\link[lmtest:gqtest]{lmtest::gqtest()}}.}

\item{.alpha}{(Optional) Critical p-value used to determine test conclusion.
The default is 0.05 (5\%).}
}
\value{
A \link[tibble:tibble-package]{tibble} with columns \code{test}, \code{statistic},
\code{p_value}, \code{result}, \code{outcome}, and \code{.notes} with extra information.
}
\description{
A wrapper around \code{\link[lmtest:gqtest]{lmtest::gqtest()}} that standardizes the inputs and outputs.
}
\details{
The hypotheses for this test are:
\itemize{
\item Null: Variances are Equal (Homoscedastic)
\item Alternative: Variances are Not Equal (Heteroscedastic)
}
}
\examples{
library(tidytest)

#> `lm` Method
mod_lm_fit <- lm(mpg ~ disp + wt + hp, data = mtcars)

goldfeld_quandt_test(mod_lm_fit)

#> `linear_reg` Method
library(parsnip)

mod_linreg_fit <- linear_reg() \%>\%
  set_engine("lm") \%>\%
  fit(mpg ~ disp + wt + hp, data = mtcars)

goldfeld_quandt_test(mod_linreg_fit)
}
\seealso{
Other homoscedasticity tests: 
\code{\link{bruesch_pagan_test}()}
}
\concept{homoscedasticity tests}
