% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_summary.R
\name{group_summary}
\alias{group_summary}
\title{Title
Summarize by group}
\usage{
group_summary(df, x, ...)
}
\arguments{
\item{df}{dataframe}

\item{x}{variable to take the min and max of}

\item{grp}{variable to group by}
}
\value{
a tibble
}
\description{
Title
Summarize by group
}
\examples{
group_summary(mtcars, mpg, cyl)
}
