% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg-linearity.R
\name{harvey_collier_test}
\alias{harvey_collier_test}
\alias{harvey_collier_test.lm}
\alias{harvey_collier_test._lm}
\alias{harvey_collier_test._glm}
\title{Run a Harvey-Collier Test}
\usage{
harvey_collier_test(object, ..., .alpha = 0.05)

\method{harvey_collier_test}{lm}(object, ..., .alpha = 0.05)

\method{harvey_collier_test}{`_lm`}(object, ..., .alpha = 0.05)

\method{harvey_collier_test}{`_glm`}(object, ..., .alpha = 0.05)
}
\arguments{
\item{object}{A model object (such as a fitted \code{\link[stats:lm]{lm}} object).}

\item{...}{Further arguments passed to \code{\link[lmtest:harvtest]{lmtest::harvtest()}}.}

\item{.alpha}{(Optional) Critical p-value used to determine test conclusion.
The default is 0.05 (5\%).}
}
\value{
A \link[tibble:tibble-package]{tibble} with columns \code{test}, \code{statistic},
\code{p_value}, \code{result}, \code{outcome}, and \code{.notes} with extra information.
}
\description{
A wrapper around \code{\link[lmtest:harvtest]{lmtest::harvtest()}} that standardizes the inputs and outputs.
}
\details{
The hypotheses for this test are:
\itemize{
\item Null: True Relationship is Linear
\item Alternative: True Relationship is Not Linear (Convex or Concave)
}
}
\examples{
library(tidytest)

#> `lm` Method
mod_lm_fit <- lm(mpg ~ disp + wt + hp, data = mtcars)

harvey_collier_test(mod_lm_fit)

#> `linear_reg` Method
library(parsnip)

mod_linreg_fit <- linear_reg() \%>\%
  set_engine("lm") \%>\%
  fit(mpg ~ disp + wt + hp, data = mtcars)

harvey_collier_test(mod_linreg_fit)
}
\seealso{
Other linearity tests: 
\code{\link{ramsey_reset_test}()}
}
\concept{linearity tests}
