% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexbin.R
\name{hexbin}
\alias{hexbin}
\title{Hex plot}
\usage{
hexbin(df, x, y, bins = 10)
}
\arguments{
\item{df}{dataframe}

\item{x}{x var}

\item{y}{y var}

\item{bins}{number of bins}
}
\value{
a hex plot
}
\description{
Hex plot
}
\examples{
hexbin(ggplot2::diamonds, carat, price)
}
