% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg-observations.R
\name{identify_extreme_leverages}
\alias{identify_extreme_leverages}
\title{Identify Extreme Leverage Points}
\usage{
identify_extreme_leverages(object, id = NULL, .multiplier = 3)
}
\arguments{
\item{object}{A model object (such as a fitted \code{\link[stats:lm]{lm}} object).}

\item{id}{(Optional) A vector of values, the same length as the number of
observations, used as an identifier for each data point. If left as NULL,
the row number will be added as the ID column.}

\item{.multiplier}{(Optional) Used to determine which leverages are
considered to be "extreme". The default is the rule-of-thumb 3
(see details).}
}
\value{
A \code{\link[tibble:tibble]{tibble}}.
}
\description{
A data point with extreme leverage means that it has an extreme value or
values in its predictor (x) values and/or an unusual combination of its
predictors values. If this is the case, the data point(s) is/are
influential, meaning that it has an outsized influence on a regression.
}
\details{
Extreme leverage points are defined as those values that are greater than
the ratio of the number of coefficients to the number of observations,
multiplied by some multiplier. A traditional rule-of-thumb is for the
multiplier to be three.
}
\examples{
library(dplyr)
library(parsnip)
library(tidytest)

mod_fit <- parsnip::linear_reg() \%>\%
  set_engine("lm") \%>\%
  fit(mpg ~ disp + wt + hp, data = mtcars)

identify_extreme_leverages(mod_fit)
identify_extreme_leverages(mod_fit, id = rownames(mtcars))

}
\references{
Kutner, M., Nachtsheim, C., Neter, J. and Li, W. (2005).
\emph{Applied Linear Statistical Models}. ISBN: 0-07-238688-6.
McGraw-Hill/Irwin.
}
