% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg-multicollinearity.R
\name{identify_multicollinearity}
\alias{identify_multicollinearity}
\alias{identify_multicollinearity.lm}
\alias{identify_multicollinearity._lm}
\title{Calculate Multicollinearity of Variables in a Linear Regression}
\usage{
identify_multicollinearity(object)

\method{identify_multicollinearity}{lm}(object)

\method{identify_multicollinearity}{`_lm`}(object)
}
\arguments{
\item{object}{A model object (such as a fitted \code{\link[stats:lm]{lm}} object).}
}
\value{
A \link[tibble:tibble-package]{tibble} with columns \code{test}, \code{statistic},
\code{p_value}, \code{result}, \code{outcome}, and \code{.notes} with extra information.
}
\description{
A wrapper around \code{\link[car:vif]{car::vif()}} that standardizes the inputs and outputs.
}
\examples{
library(dplyr)
library(parsnip)
library(tidytest)

#> `lm` Method
mod_lm_fit <- lm(mpg ~ disp + wt + hp, data = mtcars)

identify_multicollinearity(mod_lm_fit)

#> `linear_reg` Method
mod_linreg_fit <- linear_reg() \%>\%
  set_engine("lm") \%>\%
  fit(mpg ~ disp + wt + hp, data = mtcars)

identify_multicollinearity(mod_linreg_fit)
}
