% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg-independence.R
\name{ljung_box_test}
\alias{ljung_box_test}
\alias{ljung_box_test.default}
\alias{ljung_box_test.numeric}
\alias{ljung_box_test.lm}
\alias{ljung_box_test._lm}
\alias{ljung_box_test._glm}
\title{Run a Ljung-Box Test}
\usage{
ljung_box_test(object, ..., .alpha = 0.05)

\method{ljung_box_test}{default}(object, ...)

\method{ljung_box_test}{numeric}(x, ..., .alpha = 0.05)

\method{ljung_box_test}{lm}(object, ..., .alpha = 0.05)

\method{ljung_box_test}{`_lm`}(object, ..., .alpha = 0.05)

\method{ljung_box_test}{`_glm`}(object, ..., .alpha = 0.05)
}
\arguments{
\item{object}{A model object (such as a fitted \code{\link[stats:lm]{lm}} object).}

\item{...}{Further arguments passed to \code{\link[stats]{Box.test}}.}

\item{.alpha}{(Optional) Critical p-value used to determine test conclusion.
The default is 0.05 (5\%).}

\item{x}{For numeric method, a vector of residuals.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
A wrapper around \code{\link[stats]{Box.test}} that standardizes the inputs
and outputs.
}
\details{
The hypotheses for this test are:
\itemize{
\item Null: No Autocorrelation
\item Alternative: Autocorrelation
}
}
\examples{
library(dplyr)
library(parsnip)
library(tidytest)

#> Numeric Method
set.seed(1914)
resids <- rnorm(n = 100)

ljung_box_test(resids)

#> `lm` Method
mod_lm_fit <- lm(mpg ~ disp + wt + hp, data = mtcars)

ljung_box_test(mod_lm_fit)

#> Tidymodels Method
mod_linreg_fit <- parsnip::linear_reg() \%>\%
  set_engine("lm") \%>\%
  fit(mpg ~ disp + wt + hp, data = mtcars)

ljung_box_test(mod_linreg_fit)

}
