% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtcars10}
\alias{mtcars10}
\title{First 10 rows of mtcars}
\format{An object of class \code{data.frame} with 10 rows and 11 columns.}
\usage{
mtcars10
}
\description{
\itemize{
\item \code{var1}: description 1
\item \code{var2}: description 2
}
}
\keyword{datasets}
