% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tseries-stationarity.R
\name{phillips_perron_test}
\alias{phillips_perron_test}
\alias{phillips_perron_test.default}
\alias{phillips_perron_test.numeric}
\alias{phillips_perron_test.lm}
\alias{phillips_perron_test._lm}
\title{Run a Phillips-Perron Unit Root Test}
\usage{
phillips_perron_test(x, ...)

\method{phillips_perron_test}{default}(x, ...)

\method{phillips_perron_test}{numeric}(x, alternative = "stationary", ..., .alpha = 0.05)

\method{phillips_perron_test}{lm}(x, alternative = "stationary", ..., .alpha = 0.05)

\method{phillips_perron_test}{`_lm`}(x, alternative = "stationary", ..., .alpha = 0.05)
}
\arguments{
\item{x}{a  numeric vector or univariate time series (for numeric method),
or a model object (such as a fitted \code{lm} object).}

\item{...}{further arguments passed to \code{\link[tseries]{pp.test}}.}

\item{alternative}{the alternative hypothesis, either "stationary" (default)
or "explosive".}

\item{.alpha}{critical p-value used to determine test conclusion.}
}
\value{
a \link[tibble:tibble-package]{tibble}.
}
\description{
A wrapper around \code{\link[tseries]{pp.test}} that standardizes the
inputs and outputs.
}
\details{
The hypotheses for this test are:
\itemize{
\item Null: Series has a unit root
\item Alternative: Series is either stationary or has explosive root
(user defined)
}
}
\examples{
library(dplyr)
library(parsnip)
library(tidytest)

mod_fit <- parsnip::linear_reg() \%>\%
  set_engine("lm") \%>\%
  fit(mpg ~ disp + wt + hp, data = mtcars)

phillips_perron_test(mod_fit)
phillips_perron_test(mod_fit, alternative = "stationary")
phillips_perron_test(mod_fit, alternative = "explosive")

}
