% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg-plots.R
\name{plot_pred_vs_resid}
\alias{plot_pred_vs_resid}
\title{Create a Plot of Predictors vs Residuals}
\usage{
plot_pred_vs_resid(object, data, .origin = "dashed")
}
\arguments{
\item{object}{A model object (such as a fitted \code{\link[stats:lm]{lm}} object).}

\item{data}{A data set (\code{\link[base:data.frame]{data.frame}} or
\code{\link[tibble:tibble]{tibble}}) used to calculate predicted values and
residuals (likely want to use the data off of which the model was built).}

\item{.origin}{(Optional) The \code{linetype} of the line drawn at the \code{y = 0}.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
A predictors vs residuals plot is useful for visually inspecting if a linear
regression model violates both the Linear and Equal Variance assumptions
of its residuals. The plot plots predicted values on the x-axis and
residuals on the y-axis. If the resulting scatterplot shows a discernible
pattern, then either or both of the assumptions are likely violated.
}
\examples{
library(dplyr)
library(parsnip)
library(tidytest)

mod_fit <- parsnip::linear_reg() \%>\%
  set_engine("lm") \%>\%
  fit(mpg ~ disp + wt + hp, data = mtcars)

plot_pred_vs_resid(mod_fit, mtcars)

}
\references{
Kutner, M., Nachtsheim, C., Neter, J. and Li, W. (2005).
\emph{Applied Linear Statistical Models}. ISBN: 0-07-238688-6.
McGraw-Hill/Irwin.
}
