% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg-plots.R
\name{plot_predictions_vs_residuals}
\alias{plot_predictions_vs_residuals}
\alias{plot_predictions_vs_residuals.data.frame}
\title{Create a Plot of Predictions vs Residuals}
\usage{
plot_predictions_vs_residuals(.data, .hline = "dashed")

\method{plot_predictions_vs_residuals}{data.frame}(.data, .hline = "dashed")
}
\arguments{
\item{.data}{The output of \code{\link[=calculate_residuals]{calculate_residuals()}}.}

\item{.hline}{(Optional) The \code{linetype} of the line drawn at \code{y = 0}.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
A predictors vs residuals plot is useful for visually inspecting if a linear
regression model violates both the Linear and Equal Variance assumptions
of its residuals. The plot plots predicted values on the x-axis and
residuals on the y-axis. If the resulting scatterplot shows a discernible
pattern, then either or both of the assumptions are likely violated.
}
\examples{
library(tidytest)

mod_lm_fit <- lm(mpg ~ disp + wt + hp, data = mtcars)

pred_vs_resid_tbl <- calculate_residuals(mod_lm_fit)

plot_predictions_vs_residuals(pred_vs_resid_tbl)

}
\references{
Kutner, M., Nachtsheim, C., Neter, J. and Li, W. (2005).
\emph{Applied Linear Statistical Models}. ISBN: 0-07-238688-6.
McGraw-Hill/Irwin.
}
\seealso{
\code{\link[=calculate_residuals]{calculate_residuals()}}
}
