% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg-plots.R
\name{plot_qq_normality}
\alias{plot_qq_normality}
\alias{plot_qq_normality.data.frame}
\title{Create a Quantile-Quantile (Q-Q) Plot}
\usage{
plot_qq_normality(.data)

\method{plot_qq_normality}{data.frame}(.data)
}
\arguments{
\item{.data}{The output of \code{\link[=calculate_residuals]{calculate_residuals()}}.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
A QQ plot is useful for visually inspecting if a linear regression model
violates the Normality assumption of its residuals. The plot plots the
theoretical quantiles of a standard normal distribution on the x-axis and
the sample quantiles of the residuals on the y-axis. If more than just a few
of the points fall outside of the confidence band, the assumption is likely
violated.
}
\examples{
library(tidytest)

mod_lm_fit <- lm(mpg ~ disp + wt + hp, data = mtcars)

pred_vs_resid_tbl <- calculate_residuals(mod_lm_fit)

plot_qq_normality(pred_vs_resid_tbl)

}
\references{
Kutner, M., Nachtsheim, C., Neter, J. and Li, W. (2005).
\emph{Applied Linear Statistical Models}. ISBN: 0-07-238688-6.
McGraw-Hill/Irwin.
}
\seealso{
\code{\link[=calculate_residuals]{calculate_residuals()}}
}
