% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg-linearity.R
\name{ramsey_reset}
\alias{ramsey_reset}
\title{Run a Ramsey's Regression Equation Specification Error Test (RESET)}
\usage{
ramsey_reset(object, power = 2:3, ..., .alpha = 0.05)
}
\arguments{
\item{object}{a model object (such as a fitted \code{lm} object).}

\item{power}{integers. A vector of positive integers indicating
    the powers of the variables that should be included. By default,
    the test is for quadratic or cubic influence of the fitted response.}

\item{...}{further arguments passed to \code{\link[lmtest]{resettest}}.}

\item{.alpha}{critical p-value used to determine test conclusion.}
}
\value{
a \link[tibble:tibble-package]{tibble}.
}
\description{
A wrapper around \code{\link[lmtest]{resettest}} that standardizes the
inputs and outputs.
}
\details{
The hypotheses for this test are:
\itemize{
\item Null: Linear Specification is Valid
\item Alternative: Linear Specification is Not Valid
}
}
\examples{
library(dplyr)
library(parsnip)
library(tidytest)

mod_fit <- parsnip::linear_reg() \%>\%
  set_engine("lm") \%>\%
  fit(mpg ~ disp + wt + hp, data = mtcars)

ramsey_reset(mod_fit)

}
