% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytest-package.R
\docType{package}
\name{tidytest-package}
\alias{tidytest}
\alias{tidytest-package}
\title{tidytest: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
