% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg-normality.R
\name{anderson_darling_test}
\alias{anderson_darling_test}
\title{Run an Anderson-Darling Test}
\usage{
anderson_darling_test(object, .alpha = 0.05)
}
\arguments{
\item{object}{a model object (such as a fitted \code{lm} object).}

\item{.alpha}{critical p-value used to determine test conclusion.}
}
\value{
a \link[tibble:tibble-package]{tibble}.
}
\description{
A wrapper around \code{\link[nortest]{ad.test}} that standardizes the
inputs and outputs.
}
\details{
The hypotheses for this test are:
\itemize{
\item Null: Follows a Normal Distribution
\item Alternative: Does Not Follow a Normal Distribution
}
}
\examples{
library(dplyr)
library(parsnip)
library(tidytest)

mod_fit <- parsnip::linear_reg() \%>\%
  set_engine("lm") \%>\%
  fit(mpg ~ disp + wt + hp, data = mtcars)

anderson_darling_test(mod_fit)

}
