% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg-normality.R
\name{anderson_darling_test}
\alias{anderson_darling_test}
\alias{anderson_darling_test.default}
\alias{anderson_darling_test.numeric}
\alias{anderson_darling_test.lm}
\alias{anderson_darling_test._lm}
\alias{anderson_darling_test._glm}
\title{Run an Anderson-Darling Test}
\usage{
anderson_darling_test(object, ..., .alpha = 0.05)

\method{anderson_darling_test}{default}(object, ...)

\method{anderson_darling_test}{numeric}(x, ..., .alpha = 0.05)

\method{anderson_darling_test}{lm}(object, ..., .alpha = 0.05)

\method{anderson_darling_test}{`_lm`}(object, ..., .alpha = 0.05)

\method{anderson_darling_test}{`_glm`}(object, ..., .alpha = 0.05)
}
\arguments{
\item{object}{A model object (such as a fitted \code{\link[stats:lm]{lm}} object).}

\item{...}{Not currently used.}

\item{.alpha}{(Optional) Critical p-value used to determine test conclusion.
The default is 0.05 (5\%).}

\item{x}{For numeric method, a vector of residuals.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
A wrapper around \code{\link[nortest]{ad.test}} that standardizes the
inputs and outputs.
}
\details{
The hypotheses for this test are:
\itemize{
\item Null: Follows a Normal Distribution
\item Alternative: Does Not Follow a Normal Distribution
}
}
\examples{
library(dplyr)
library(parsnip)
library(tidytest)

#> Numeric Method
set.seed(1914)
resids <- rnorm(n = 100)

anderson_darling_test(resids)

#> `lm` Method
mod_lm_fit <- lm(mpg ~ disp + wt + hp, data = mtcars)

anderson_darling_test(mod_lm_fit)

#> Tidymodels Method
mod_linreg_fit <- parsnip::linear_reg() \%>\%
  set_engine("lm") \%>\%
  fit(mpg ~ disp + wt + hp, data = mtcars)

anderson_darling_test(mod_linreg_fit)

}
