% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg-homoscedasticity.R
\name{bruesch_pagan_test}
\alias{bruesch_pagan_test}
\title{Run a Bruesch-Pagan Test}
\usage{
bruesch_pagan_test(object, ..., .alpha = 0.05)
}
\arguments{
\item{object}{a model object (such as a fitted \code{lm} object).}

\item{...}{further arguments passed to \code{\link[lmtest]{bptest}}.}

\item{.alpha}{critical p-value used to determine test conclusion.}
}
\value{
a \link[tibble:tibble-package]{tibble}.
}
\description{
A wrapper around \code{\link[lmtest]{bptest}} that standardizes the inputs
and outputs.
}
\examples{
library(dplyr)
library(parsnip)
library(tidytest)

mod_fit <- parsnip::linear_reg() \%>\%
  set_engine("lm") \%>\%
  fit(mpg ~ disp + wt + hp, data = mtcars)

bruesch_pagan_test(mod_fit)

}
