% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyl_plot.R
\name{cyl_plot}
\alias{cyl_plot}
\title{Create a bar plot of the mtcars cylinder column}
\usage{
cyl_plot()
}
\value{
a bar plot
}
\description{
Create a bar plot of the mtcars cylinder column
}
\examples{
cyl_plot()
}
