% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyl_sum.R
\name{cyl_sum}
\alias{cyl_sum}
\title{Summarize the mtcars mpg & cylinder columns}
\usage{
cyl_sum()
}
\value{
a table summarizing columns 'cyl' and 'mpg'
}
\description{
Summarize the mtcars mpg & cylinder columns
}
