% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg-homoscedasticity.R
\name{goldfeld_quandt_test}
\alias{goldfeld_quandt_test}
\alias{goldfeld_quandt_test.default}
\alias{goldfeld_quandt_test.lm}
\alias{goldfeld_quandt_test._lm}
\alias{goldfeld_quandt_test._glm}
\title{Run a Goldfeld-Quandt Test}
\usage{
goldfeld_quandt_test(object, alternative = "two.sided", ..., .alpha = 0.05)

\method{goldfeld_quandt_test}{default}(object, ...)

\method{goldfeld_quandt_test}{lm}(object, alternative = "two.sided", ..., .alpha = 0.05)

\method{goldfeld_quandt_test}{`_lm`}(object, alternative = "two.sided", ..., .alpha = 0.05)

\method{goldfeld_quandt_test}{`_glm`}(object, alternative = "two.sided", ..., .alpha = 0.05)
}
\arguments{
\item{object}{A model object (such as a fitted \code{\link[stats:lm]{lm}} object).}

\item{alternative}{The alternative hypothesis, one of "two.sided" (default),
"greater", or "less".}

\item{...}{Further arguments passed to \code{\link[lmtest]{gqtest}}.}

\item{.alpha}{(Optional) Critical p-value used to determine test conclusion.
The default is 0.05 (5\%).}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
A wrapper around \code{\link[lmtest]{gqtest}} that standardizes the inputs
and outputs.
}
\details{
The hypotheses for this test are:
\itemize{
\item Null: Variances are Equal (Homoscedastic)
\item Alternative: Variances are Not Equal (Heteroscedastic)
}
}
\examples{
library(dplyr)
library(parsnip)
library(tidytest)

mod_fit <- parsnip::linear_reg() \%>\%
  set_engine("lm") \%>\%
  fit(mpg ~ disp + wt + hp, data = mtcars)

goldfeld_quandt_test(mod_fit)

}
