% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg-homoscedasticity.R
\name{goldfeld_quandt_test}
\alias{goldfeld_quandt_test}
\title{Run a Goldfeld-Quandt Test}
\usage{
goldfeld_quandt_test(object, alternative = "two.sided", ..., .alpha = 0.05)
}
\arguments{
\item{object}{a model object (such as a fitted \code{lm} object).}

\item{alternative}{a character string specifying the alternative hypothesis.
    The default is to test for increasing variances.}

\item{...}{further arguments passed to \code{\link[lmtest]{gqtest}}.}

\item{.alpha}{critical p-value used to determine test conclusion.}
}
\value{
a \link[tibble:tibble-package]{tibble}.
}
\description{
A wrapper around \code{\link[lmtest]{gqtest}} that standardizes the inputs
and outputs.
}
\details{
The hypotheses for this test are:
\itemize{
\item Null: Variances are Equal (Homoscedastic)
\item Alternative: Variances are Not Equal (Heteroscedastic)
}
}
\examples{
library(dplyr)
library(parsnip)
library(tidytest)

mod_fit <- parsnip::linear_reg() \%>\%
  set_engine("lm") \%>\%
  fit(mpg ~ disp + wt + hp, data = mtcars)

goldfeld_quandt_test(mod_fit)

}
