% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg-linearity.R
\name{harvey_collier_test}
\alias{harvey_collier_test}
\alias{harvey_collier_test.default}
\alias{harvey_collier_test.lm}
\alias{harvey_collier_test._lm}
\alias{harvey_collier_test._glm}
\title{Run a Harvey-Collier Test}
\usage{
harvey_collier_test(object, ..., .alpha = 0.05)

\method{harvey_collier_test}{default}(object, ...)

\method{harvey_collier_test}{lm}(object, ..., .alpha = 0.05)

\method{harvey_collier_test}{`_lm`}(object, ..., .alpha = 0.05)

\method{harvey_collier_test}{`_glm`}(object, ..., .alpha = 0.05)
}
\arguments{
\item{object}{A model object (such as a fitted \code{\link[stats:lm]{lm}} object).}

\item{...}{Further arguments passed to \code{\link[lmtest]{harvtest}}.}

\item{.alpha}{(Optional) Critical p-value used to determine test conclusion.
The default is 0.05 (5\%).}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
A wrapper around \code{\link[lmtest]{harvtest}} that standardizes the
inputs and outputs.
}
\details{
The hypotheses for this test are:
\itemize{
\item Null: True Relationship is Linear
\item Alternative: True Relationship is Not Linear (Convex or Concave)
}
}
\examples{
library(dplyr)
library(parsnip)
library(tidytest)

mod_fit <- parsnip::linear_reg() \%>\%
  set_engine("lm") \%>\%
  fit(mpg ~ disp + wt + hp, data = mtcars)

harvey_collier_test(mod_fit)

}
