% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg-observations.R
\name{identify_outliers}
\alias{identify_outliers}
\alias{identify_outliers.lm}
\title{Identify Outliers}
\usage{
identify_outliers(object, id = NULL, .cutoff = 3)

\method{identify_outliers}{lm}(object, id = NULL, .cutoff = 3)
}
\arguments{
\item{object}{A model object (such as a fitted \code{\link[stats:lm]{lm}} object).}

\item{id}{(Optional) A vector of values, the same length as the number of
observations, used as an identifier for each data point. If left as NULL,
the row number will be added as the ID column.}

\item{.cutoff}{(Optional) Used to determine which standard residuals are
indicative of an outlier. The default is the rule-of-thumb 3
(see details).}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
A data point flagged as an outlier means that is has an extreme value in its
response (y) variable. If this is the case, the data point(s) is/are
influential, meaning that it has an outsized influence on a regression.
}
\details{
Outliers are defined as those data points that have a standardized
residual value greater than some cutoff value. A traditional rule-of-thumb
is for that cutoff value to be three.
}
\examples{
library(tidytest)

#> `lm` Method
mod_lm_fit <- lm(mpg ~ disp + wt + hp, data = mtcars)

identify_outliers(mod_lm_fit)
identify_outliers(mod_lm_fit, id = rownames(mtcars))
}
\references{
Kutner, M., Nachtsheim, C., Neter, J. and Li, W. (2005).
\emph{Applied Linear Statistical Models}. ISBN: 0-07-238688-6.
McGraw-Hill/Irwin.
}
