% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tseries-stationarity.R
\name{kpss_test}
\alias{kpss_test}
\alias{kpss_test.default}
\alias{kpss_test.numeric}
\alias{kpss_test.lm}
\alias{kpss_test._lm}
\title{Run a Kwiatkowski-Phillips-Schmidt-Shin (KPSS) Test}
\usage{
kpss_test(x, ...)

\method{kpss_test}{default}(x, ...)

\method{kpss_test}{numeric}(x, null = "Level", ..., .alpha = 0.05)

\method{kpss_test}{lm}(x, null = "Level", ..., .alpha = 0.05)

\method{kpss_test}{`_lm`}(x, null = "Level", ..., .alpha = 0.05)
}
\arguments{
\item{x}{a numeric vector or univariate time series (for numeric method),
or a model object (such as a fitted \code{lm} object).}

\item{...}{further arguments passed to \code{\link[tseries]{kpss.test}}.}

\item{null}{the null hypothesis, either "Level" (default) or "Trend".}

\item{.alpha}{critical p-value used to determine test conclusion.}
}
\value{
a \link[tibble:tibble-package]{tibble}.
}
\description{
A wrapper around \code{\link[tseries]{kpss.test}} that standardizes the
inputs and outputs.
}
\details{
The hypotheses for this test are:
\itemize{
\item Null: Series is either Level or Trend stationary (user defined)
\item Alternative: Series has a unit root
}
}
\examples{
library(dplyr)
library(parsnip)
library(tidytest)

mod_fit <- parsnip::linear_reg() \%>\%
  set_engine("lm") \%>\%
  fit(mpg ~ disp + wt + hp, data = mtcars)

kpss_test(mod_fit)
kpss_test(mod_fit, null = "Level")
kpss_test(mod_fit, null = "Trend")

}
