% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stationarity.R
\name{kpss_test}
\alias{kpss_test}
\title{Run a Kwiatkowski-Phillips-Schmidt-Shin (KPSS) Test}
\usage{
kpss_test(x, null = "Level", ..., .alpha = 0.05)
}
\arguments{
\item{x}{a numeric vector or univariate time series.}

\item{null}{indicates the null hypothesis and must be one of
    \code{"Level"} (default) or \code{"Trend"}.  You can specify just
    the initial letter.}

\item{...}{further arguments passed to \code{\link[tseries]{adf.test}}.}

\item{.alpha}{critical p-value used to determine test conclusion.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
A wrapper around \code{\link[tseries]{kpss.test}} that standardizes the
inputs and outputs.
}
\examples{
set.seed(123)
x <- runif(n = 1000)

kpss_test(x)
kpss_test(x, null = "Trend")

}
