% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg-independence.R
\name{ljung_box_test}
\alias{ljung_box_test}
\title{Run a Ljung-Box Test}
\usage{
ljung_box_test(object, ..., .alpha = 0.05)
}
\arguments{
\item{object}{a model object (such as a fitted \code{lm} object).}

\item{...}{further arguments passed to \code{\link[stats]{Box.test}}.}

\item{.alpha}{critical p-value used to determine test conclusion.}
}
\value{
a \link[tibble:tibble-package]{tibble}.
}
\description{
A wrapper around \code{\link[stats]{Box.test}} that standardizes the inputs
and outputs.
}
\details{
The hypotheses for this test are:
\itemize{
\item Null: No Autocorrelation
\item Alternative: Autocorrelation
}
}
\examples{
library(dplyr)
library(parsnip)
library(tidytest)

mod_fit <- parsnip::linear_reg() \%>\%
  set_engine("lm") \%>\%
  fit(mpg ~ disp + wt + hp, data = mtcars)

ljung_box_test(mod_fit)

}
