% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg-independence.R
\name{ljung_box_test}
\alias{ljung_box_test}
\alias{ljung_box_test.lm}
\alias{ljung_box_test._lm}
\alias{ljung_box_test._glm}
\alias{ljung_box_test_vec}
\title{Run a Ljung-Box Test}
\usage{
ljung_box_test(object, ..., .alpha = 0.05)

\method{ljung_box_test}{lm}(object, ..., .alpha = 0.05)

\method{ljung_box_test}{`_lm`}(object, ..., .alpha = 0.05)

\method{ljung_box_test}{`_glm`}(object, ..., .alpha = 0.05)

ljung_box_test_vec(x, ..., .alpha = 0.05)
}
\arguments{
\item{object}{A model object (such as a fitted \code{\link[stats:lm]{lm}} object).}

\item{...}{Further arguments passed to \code{\link[stats:box.test]{stats::Box.test()}}.}

\item{.alpha}{(Optional) Critical p-value used to determine test conclusion.
The default is 0.05 (5\%).}

\item{x}{For \verb{_vec()} functions, a numeric vector.}
}
\value{
A \link[tibble:tibble-package]{tibble} with columns \code{test}, \code{statistic},
\code{p_value}, \code{result}, \code{outcome}, and \code{.notes} with extra information.
}
\description{
A wrapper around \code{\link[stats:box.test]{stats::Box.test()}} that standardizes the inputs and outputs.
}
\details{
The hypotheses for this test are:
\itemize{
\item Null: No Autocorrelation
\item Alternative: Autocorrelation
}
}
\examples{
library(tidytest)

#> `lm` Method
mod_lm_fit <- lm(mpg ~ disp + wt + hp, data = mtcars)

ljung_box_test(mod_lm_fit)

#> `linear_reg` Method
library(parsnip)

mod_linreg_fit <- linear_reg() \%>\%
  set_engine("lm") \%>\%
  fit(mpg ~ disp + wt + hp, data = mtcars)

ljung_box_test(mod_linreg_fit)

#> Numeric Method
set.seed(1914)
resids <- rnorm(100)

ljung_box_test_vec(resids)
}
\seealso{
Other independence tests: 
\code{\link{durbin_watson_test}()}
}
\concept{independence tests}
