% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg-plots.R
\name{plot_pred_vs_resid}
\alias{plot_pred_vs_resid}
\alias{plot_pred_vs_resid.lm}
\title{Create a Plot of Predictors vs Residuals}
\usage{
plot_pred_vs_resid(object, .hline = "dashed")

\method{plot_pred_vs_resid}{lm}(object, .hline = "dashed")
}
\arguments{
\item{object}{A model object (such as a fitted \code{\link[stats:lm]{lm}} object).}

\item{.hline}{(Optional) The \code{linetype} of the line drawn at the \code{y = 0}.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
A predictors vs residuals plot is useful for visually inspecting if a linear
regression model violates both the Linear and Equal Variance assumptions
of its residuals. The plot plots predicted values on the x-axis and
residuals on the y-axis. If the resulting scatterplot shows a discernible
pattern, then either or both of the assumptions are likely violated.
}
\examples{
library(dplyr)
library(parsnip)
library(tidytest)

#> `lm` Method
mod_lm_fit <- lm(mpg ~ disp + wt + hp, data = mtcars)

plot_pred_vs_resid(mod_lm_fit)

}
\references{
Kutner, M., Nachtsheim, C., Neter, J. and Li, W. (2005).
\emph{Applied Linear Statistical Models}. ISBN: 0-07-238688-6.
McGraw-Hill/Irwin.
}
