% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg-plots.R
\name{plot_qq_norm}
\alias{plot_qq_norm}
\title{Create a Quantile-Quantile (Q-Q) Plot}
\usage{
plot_qq_norm(object, data)
}
\arguments{
\item{object}{A model object (such as a fitted \code{\link[stats:lm]{lm}} object).}

\item{data}{A data set (\code{\link[base:data.frame]{data.frame}} or
\code{\link[tibble:tibble]{tibble}}) used to calculate predicted values and
residuals (likely want to use the data off of which the model was built).}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
A QQ plot is useful for visually inspecting if a linear regression model
violates the Normality assumption of its residuals. The plot plots the
theoretical quantiles of a standard normal distribution on the x-axis and
the sample quantiles of the residuals on the y-axis. If more than just a few
of the points fall outside of the confidence band, the assumption is likely
violated.
}
\examples{
library(dplyr)
library(parsnip)
library(tidytest)

mod_fit <- parsnip::linear_reg() \%>\%
  set_engine("lm") \%>\%
  fit(mpg ~ disp + wt + hp, data = mtcars)

plot_qq_norm(mod_fit, mtcars)

}
\references{
Kutner, M., Nachtsheim, C., Neter, J. and Li, W. (2005).
\emph{Applied Linear Statistical Models}. ISBN: 0-07-238688-6.
McGraw-Hill/Irwin.
}
