% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg-normality.R
\name{shapiro_wilk_test}
\alias{shapiro_wilk_test}
\title{Run a Shapiro-Wilk Test}
\usage{
shapiro_wilk_test(object, ..., .alpha = 0.05)
}
\arguments{
\item{object}{a model object (such as a fitted \code{lm} object).}

\item{...}{not currently used.}

\item{.alpha}{critical p-value used to determine test conclusion.}
}
\value{
a \link[tibble:tibble-package]{tibble}.
}
\description{
A wrapper around \code{\link[stats]{shapiro.test}} that standardizes the
inputs and outputs.
}
\details{
The hypotheses for this test are:
\itemize{
\item Null: Follows a Normal Distribution
\item Alternative: Does Not Follow a Normal Distribution
}
}
\examples{
library(dplyr)
library(parsnip)
library(tidytest)

mod_fit <- parsnip::linear_reg() \%>\%
  set_engine("lm") \%>\%
  fit(mpg ~ disp + wt + hp, data = mtcars)

shapiro_wilk_test(mod_fit)

}
