% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytest-package.R
\docType{package}
\name{tidytest-package}
\alias{tidytest}
\alias{tidytest-package}
\title{tidytest: Tidy Statistical Modeling Tests}
\description{
The objective of this package is to streamline the implementation of statistical tests usually used with statistical models that coheres with the tidy design framework.
}
\author{
\strong{Maintainer}: Robert Lankford \email{robbylankford@gmail.com}

}
\keyword{internal}
